Tries pulling container from DockerHub.
Also relevant to #10. This approach was timed to take 1m19s.
Theoretically, it has a limit of 200 pulls per 6 hours as long as
authentication is working correctly. Open a new issue if you start
bumping against these limits. Also, downside to this approach is that
you have to trigger a new build everytime dev changes.... Buuut it might
be possible to set up a GitHub Action to do that build and push for you!